% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LikeThis-package.R
\name{LikeThis-package}
\alias{LikeThis}
\alias{LikeThis-package}
\title{Adds infix regular expression operators to R.}
\description{
R package \bold{LikeThis} adds infix regular expression operators to R that will be
familiar to perl users. The package provides two matching operators: \code{\%~\%},
which returns a MatchResultR (which has methods for accessing capture groups), and
\code{\%?~\%}, which returns a logical and makes capture groups accessible in the local
environment via the \code{.match} list variable. The later can be used iteratively
in a loop to perform progressive matching. The package also provides a substitution
operator, \code{\%!~\%}.
}
\seealso{
\code{\link{RegularExpression}}, \code{\link{RegExpR-class}}
}

