% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LikeThis.R
\name{\%~\%}
\alias{\%~\%}
\title{Infix regular expression match operator.}
\usage{
lhs \%~\% rhs
}
\arguments{
\item{lhs}{the string to be matched}

\item{rhs}{a regular expression string or RegularExpression
object (obtained from `get.regexp` or `RE`).}
}
\value{
a `MatchResult` if there was a match, otherwise NULL.
}
\description{
Note: if lhs is not a string, this method tries to coerce
it to one.
}

