% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegExpR.R
\docType{class}
\name{MatchResultR-class}
\alias{MatchResultR}
\alias{MatchResultR-class}
\title{MatchResult object encapsulating regexec match results.}
\description{
MatchResult object encapsulating regexec match results.
}

