% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LikeThis.R
\name{RE}
\alias{RE}
\title{Convenience function for passing a pattern and flags to the right-hand side
of an infix regular expression operator.}
\usage{
RE(pattern, ...)
}
\arguments{
\item{pattern}{}

\item{...}{flag names (can be single-character)}
}
\value{
RegularExpression
}
\description{
Convenience function for passing a pattern and flags to the right-hand side
of an infix regular expression operator.
}
\examples{
# matches because the 'i' flag makes it case-insensitive
# and the 's' flag matches the dot to newlines
'a\\nb' \%~\% RE('A.B', 'i', 's')
}

