% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegExpR.R
\docType{class}
\name{RegExpR-class}
\alias{RegExpR}
\alias{RegExpR-class}
\title{RegExp class that uses native R regular expresssion functions.
Any alternative implementations should implement the same API.}
\description{
RegExp class that uses native R regular expresssion functions.
Any alternative implementations should implement the same API.
}

