% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LikeThis.R
\docType{class}
\name{RegularExpression-class}
\alias{RegularExpression}
\alias{RegularExpression-class}
\title{Class that encapsulates an re2 regexp and provides matching.}
\description{
Class that encapsulates an re2 regexp and provides matching.
}

