% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LikeThis.R
\name{get.regexp}
\alias{get.regexp}
\title{Fetch a cached regexp for `pattern` if it was
previously requested, or compile the pattern and cache the
regexp.}
\usage{
get.regexp(pattern, recompile = FALSE)
}
\arguments{
\item{pattern}{character vector of POSIX-compatible regular
expressions.}

\item{recompile}{ignore any cached regexp for `pattern`}
}
\value{
`RegularExpression` object (regexp)
}
\description{
Fetch a cached regexp for `pattern` if it was
previously requested, or compile the pattern and cache the
regexp.
}

