% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LikeThis.R
\docType{class}
\name{RegularExpression-class}
\alias{RegularExpression}
\alias{RegularExpression-class}
\title{Class that encapsulates a RegExp and provides matching. This class shouldn't
be instantiated directly; instead use `get.regexp` or one of the infix
operators.}
\description{
Class that encapsulates a RegExp and provides matching. This class shouldn't
be instantiated directly; instead use `get.regexp` or one of the infix
operators.
}

